requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-multipart
pymupdf
chromadb
pytest
httpx
